# backend/swagger/generated_swagger/create_gmail_api_o_auth2_token_request.go
// Code generated by go-swagger; DO NOT EDIT.

package generated_swagger

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// CreateGmailAPIOAuth2TokenRequest CreateGmailApiOauth2TokenRequest
//
// swagger:model CreateGmailApiOAuth2TokenRequest
type CreateGmailAPIOAuth2TokenRequest struct {

	// code
	// Required: true
	Code *string `json:"code"`

	// user id
	// Required: true
	UserID *int64 `json:"user_id"`
}

// Validate validates this create gmail Api o auth2 token request
func (m *CreateGmailAPIOAuth2TokenRequest) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateCode(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateUserID(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *CreateGmailAPIOAuth2TokenRequest) validateCode(formats strfmt.Registry) error {

	if err := validate.Required("code", "body", m.Code); err != nil {
		return err
	}

	return nil
}

func (m *CreateGmailAPIOAuth2TokenRequest) validateUserID(formats strfmt.Registry) error {

	if err := validate.Required("user_id", "body", m.UserID); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this create gmail Api o auth2 token request based on context it is used
func (m *CreateGmailAPIOAuth2TokenRequest) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CreateGmailAPIOAuth2TokenRequest) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CreateGmailAPIOAuth2TokenRequest) UnmarshalBinary(b []byte) error {
	var res CreateGmailAPIOAuth2TokenRequest
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}