# backend/swagger/generated_swagger/create_gmail_api_o_auth2_token_response.go
// Code generated by go-swagger; DO NOT EDIT.

package generated_swagger

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// CreateGmailAPIOAuth2TokenResponse CreateGmailApiOAuth2TokenResponse
//
// swagger:model CreateGmailApiOAuth2TokenResponse
type CreateGmailAPIOAuth2TokenResponse struct {

	// id
	ID int64 `json:"id,omitempty"`
}

// Validate validates this create gmail Api o auth2 token response
func (m *CreateGmailAPIOAuth2TokenResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this create gmail Api o auth2 token response based on context it is used
func (m *CreateGmailAPIOAuth2TokenResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *CreateGmailAPIOAuth2TokenResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *CreateGmailAPIOAuth2TokenResponse) UnmarshalBinary(b []byte) error {
	var res CreateGmailAPIOAuth2TokenResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}